Save 8 bytes

+++ src/lib/libast/include/ast_std.h
@@ -247,7 +247,7 @@ typedef struct
 #if !AST_NOMULTIBYTE
 	struct				/* ast.mb.* -- multibyte encoding/decoding state */
 	{
-	size_t		cur_max;	/* current maximum length in bytes of a character: > 1 == multibyte locale */
+	uint32_t	cur_max;	/* current maximum length in bytes of a character: > 1 == multibyte locale */
 	uint32_t	sync;		/* length of invalid multibyte character */
 	wchar_t		tmp_w;		/* scratch */
 	int		tmp_i;		/* scratch */
